rag/requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-openai
langchain-chroma
langchain-text-splitters
pypdf
cryptography
chromadb
openai
requests

